# docs/api/openapi.yml
openapi: 3.0.0
info:
    title: WildDuck API
    description: WildDuck API docs
    version: 1.0.0
    contact:
        url: 'https://github.com/nodemailer/wildduck'
servers:
    - url: 'https://api.wildduck.email'
...
paths:
...
    '/users/{user}/mailboxes/{mailbox}':
        delete:
            tags:
                - Mailboxes
            summary: Delete a Mailbox
            operationId: deleteMailbox
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
        get:
            tags:
                - Mailboxes
            summary: Request Mailbox information
            operationId: getMailbox
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMailboxResponse'
        put:
            tags:
                - Mailboxes
            summary: Update Mailbox information
            operationId: updateMailbox
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateMailboxRequest'
                required: true
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
        parameters:
            - name: user
              in: path
              description: Users unique ID
              required: true
              schema:
                  type: string
            - name: mailbox
              in: path
              description: Mailbox unique ID. Special use folders and INBOX can not be deleted
              required: true
              schema:
                  type: string
    '/users/{user}/mailboxes':
        get:
            tags:
                - Mailboxes
            summary: List Mailboxes for a User
            operationId: getMailboxes
            parameters:
                - name: specialUse
                  in: query
                  description: Should the response include only folders with specialUse flag set.
                  schema:
                      type: boolean
                - name: showHidden
                  in: query
                  description: Hidden folders are not included in the listing by default.
                  schema:
                      type: boolean
                - name: counters
                  in: query
                  description: Should the response include counters (total + unseen). Counters come with some overhead.
                  schema:
                      type: boolean
                - name: sizes
                  in: query
                  description: Should the response include mailbox size in bytes. Size numbers come with a lot of overhead as an aggregated query is ran.
                  schema:
                      type: boolean
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMailboxesResponse'
        post:
            tags:
                - Mailboxes
            summary: Create new Mailbox
            operationId: createMailbox
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMailboxRequest'
                required: true
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateMailboxResponse'
        parameters:
            - name: user
              in: path
              description: Users unique ID
              required: true
              schema:
                  type: string
...
components:
    schemas:
...
        SpecialUse:
            enum:
                - \Drafts
                - \Junk
                - \Sent
                - \Trash
            type: string
            description: 'Either special use identifier or null. One of \Drafts, \Junk, \Sent or \Trash'
...
        SuccessResponse:
            required:
                - success
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
...
        GetMailboxResponse:
            required:
                - success
                - id
                - name
                - path
                - specialUse
                - modifyIndex
                - subscribed
                - hidden
                - total
                - unseen
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                id:
                    type: string
                    description: Mailbox ID
                name:
                    type: string
                    description: Name for the mailbox (unicode string)
                path:
                    type: string
                    description: 'Full path of the mailbox, folders are separated by slashes, ends with the mailbox name (unicode string)'
                specialUse:
                    $ref: '#/components/schemas/SpecialUse'
                modifyIndex:
                    type: number
                    description: Modification sequence number. Incremented on every change in the mailbox.
                subscribed:
                    type: boolean
                    description: Mailbox subscription status. IMAP clients may unsubscribe from a folder.
                hidden:
                    type: boolean
                    description: Is the folder hidden or not
                total:
                    type: number
                    description: How many messages are stored in this mailbox
                unseen:
                    type: number
                    description: How many unseen messages are stored in this mailbox
        GetMailboxesResponse:
            required:
                - success
                - results
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetMailboxesResult'
                    description: List of user mailboxes
        CreateMailboxResponse:
            required:
                - success
                - id
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                id:
                    type: string
                    description: Mailbox ID
...
        UpdateMailboxRequest:
            type: object
            properties:
                path:
                    type: string
                    description: 'Full path of the mailbox, use this to rename an existing Mailbox'
                hidden:
                    type: boolean
                    description: Is the folder hidden or not. Hidden folders can not be opened in IMAP.
                retention:
                    type: number
                    description: Retention policy for the Mailbox (in ms). Changing retention value only affects messages added to this folder after the change
                subscribed:
                    type: boolean
                    description: Change Mailbox subscription state
        CreateMailboxRequest:
            required:
                - path
            type: object
            properties:
                path:
                    type: string
                    description: 'Full path of the mailbox, folders are separated by slashes, ends with the mailbox name (unicode string)'
                hidden:
                    type: boolean
                    description: Is the folder hidden or not. Hidden folders can not be opened in IMAP.
                retention:
                    type: number
                    description: Retention policy for the created Mailbox. Milliseconds after a message added to mailbox expires. Set to 0 to disable.
...
        GetMailboxesResult:
            required:
                - id
                - name
                - path
                - specialUse
                - modifyIndex
                - subscribed
                - hidden
                - total
                - unseen
            type: object
            properties:
                id:
                    type: string
                    description: Mailbox ID
                name:
                    type: string
                    description: Name for the mailbox (unicode string)
                path:
                    type: string
                    description: 'Full path of the mailbox, folders are separated by slashes, ends with the mailbox name (unicode string)'
                specialUse:
                    type: string
                    description: 'Either special use identifier or null. One of \Drafts, \Junk, \Sent or \Trash'
                modifyIndex:
                    type: number
                    description: Modification sequence number. Incremented on every change in the mailbox.
                subscribed:
                    type: boolean
                    description: Mailbox subscription status. IMAP clients may unsubscribe from a folder.
                retention:
                    type: number
                    description: Default retention policy for this mailbox (in ms). If set then messages added to this maibox will be automatically deleted after retention time.
                hidden:
                    type: boolean
                    description: Is the folder hidden or not
                total:
                    type: number
                    description: How many messages are stored in this mailbox
                unseen:
                    type: number
                    description: How many unseen messages are stored in this mailbox
...
    securitySchemes:
        AccessTokenAuth:
            name: X-Access-Token
            type: apiKey
            in: header
            description: |-
                If authentication is enabled in the WildDuck configuration, you will need to supply an access token in the `X-Access-Token` header.

                ```json
                {
                  "X-Access-Token": "59fc66a03e54454869460e45"
                }
                ```
security:
    - AccessTokenAuth: []